implement jwt guard and isAuthorOrAdmin error message

+++ src/blogpost/blogpost.controller.ts
@@ -7,41 +7,74 @@ import {
   Param,
   Delete,
   UseGuards,
+  Request,
+  NotFoundException,
+  ForbiddenException,
 } from '@nestjs/common';
+import { ISPUBLIC } from '../common/decorator';
+import { UserType } from 'src/user/enums/user-type-enum';
 import { JwtAuthGuard } from '../common/guard/jwt-auth.guard';
 import { BlogpostService } from './blogpost.service';
 import { BlogPostDto } from './dto/blogpost.dto';
 import { UpdateBlogpostDto } from './dto/update-blogpost.dto';
 
+@UseGuards(JwtAuthGuard)
 @Controller('blogpost')
 export class BlogpostController {
   constructor(private readonly blogpostService: BlogpostService) {}
 
-  @UseGuards(JwtAuthGuard)
   @Post()
   async create(@Body() blogPostDto: BlogPostDto) {
     return await this.blogpostService.create(blogPostDto);
   }
 
+  @ISPUBLIC()
   @Get()
   async findAll() {
     return await this.blogpostService.findAll();
   }
 
+  @ISPUBLIC()
   @Get(':id')
   async findOne(@Param('id') id: string) {
     return await this.blogpostService.findOne(id);
   }
 
-  @UseGuards(JwtAuthGuard)
   @Patch()
-  async update(@Body() updateBlogpostDto: UpdateBlogpostDto) {
-    return await this.blogpostService.update(updateBlogpostDto);
+  async update(@Body() updateBlogpostDto: UpdateBlogpostDto, @Request() req) {
+    const foundBlogPost = await this.findOne(updateBlogpostDto.id);
+
+    if (!foundBlogPost) throw new NotFoundException();
+
+    if (
+      foundBlogPost.userId === req.user.id ||
+      req.user.userType === UserType.ADMIN
+    )
+      return await this.blogpostService.update(updateBlogpostDto);
+
+    throw new ForbiddenException({
+      statusCode: 403,
+      message:
+        'Forbidden. You have the be the author of the blog post or an admin.',
+    });
   }
 
-  @UseGuards(JwtAuthGuard)
   @Delete(':id')
-  async remove(@Param('id') id: string) {
-    return await this.blogpostService.remove(id);
+  async remove(@Param('id') id: string, @Request() req) {
+    const foundBlogPost = await this.findOne(id);
+
+    if (!foundBlogPost) throw new NotFoundException();
+
+    if (
+      foundBlogPost.userId === req.user.id ||
+      req.user.userType === UserType.ADMIN
+    )
+      return await this.blogpostService.remove(id);
+
+    throw new ForbiddenException({
+      statusCode: 403,
+      message:
+        'Forbidden. You have the be the author of the blog post or an admin.',
+    });
   }
 }
